casting uint shuffles to int for cuda5.5 (todi)

+++ mpi-dpd/scan-mauro.cu
@@ -19,7 +19,7 @@ __global__ void breduce(uint4 *vin, unsigned int *vout, int n) {
 
 	#pragma unroll
 	for(int i = 16; i > 0; i >>= 1)
-		val.x += __shfl_down(val.x, i);
+		val.x += __shfl_down((int)val.x, i);
 
 	if (0 == lid)
 		shtmp[wid] = val.x;
@@ -30,7 +30,7 @@ __global__ void breduce(uint4 *vin, unsigned int *vout, int n) {
 		
 		#pragma unroll
 		for(int i = 16; i > 0; i >>= 1)
-			val.x += __shfl_down(val.x, i);
+			val.x += __shfl_down((int)val.x, i);
 	}
 	if (0 == threadIdx.x) vout[blockIdx.x] = val.x;
 	return;
@@ -136,7 +136,7 @@ __global__ void gexscan(uint4 *vin, unsigned int *offs, uint4 *vout, int n) {
 	tu4.w = tmp;
 	#pragma unroll
 	for(int i = 1; i < 32; i <<= 1)
-		tu4.w += (lid >= i)*__shfl_up(tu4.w, i);
+		tu4.w += (lid >= i)*__shfl_up((int)tu4.w, i);
 
 	if (lid == 31)
 		if (wid < NWARP-1) woff[wid+1] = tu4.w;
@@ -149,7 +149,7 @@ __global__ void gexscan(uint4 *vin, unsigned int *offs, uint4 *vout, int n) {
 		tmp = (lid < NWARP) ? woff[lid] : 0;
 		#pragma unroll
 		for(int i = 1; i < NWARP; i <<= 1)
-			tmp += (lid >= i)*__shfl_up(tmp, i);
+			tmp += (lid >= i)*__shfl_up((int)tmp, i);
 
 		if (lid < NWARP) woff[lid] = tmp;
 	}
